requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-dotenv
PyMuPDF
cohere
scikit-learn
requests
